+++ app/models/pipeline_step/heart_model_calculation.rb
@@ -23,6 +23,7 @@ module PipelineStep
 
     def internal_run
       computation.script = ScriptGenerator::HeartModel.new(pipeline).call
+      computation.job_id = nil
       computation.save!
 
       Rimrock::StartJob.perform_later computation